An initial implementation of the model.
This is an attempt to produce a Python implementation of the algorithm
implemented by the previous Scala-based simulation.

One should type `pip install -r requirements.txt` from the root
directory of the project.

Tests are included for the table, table lookups, and the model.  Those
tests can be run by typing `nosetests` in the root directory of the
project.

Fixes azavea/civic-apps#3