Add support for source and binary Python package distributions
This is based on the outline provided by the Python Packaging User
Guide:

  https://packaging.python.org/en/latest/distributing.html

The goal is for this work to support building and uploading package
distributions directly to PyPI from Travis CI. Adding something
like:

  distributions: "sdist bdist_wheel"

To the PyPI `deploy` provider should handle most of the process.

Also, for building local binary distributions, the `wheel` package must
be installed.