Initial steps for adding water quality
Changes the structure of the input to account for separating
modifications from the original distribution of soil/land use pairings.
Also adds some routines for calculating volume and getting pollution
load.

Add Event Mean Concentration loads for NLCD types:

This table maps an NLCD class to the Event Mean Concentration for
particular pollution types.  This data was presented as Table 1 in a
document titled "Outline of Water Quality Algorithms to be added to the
Current Runoff Model" from Barry Evans.

Event Mean Concentrations are an estimate of loads for a given land use
type:

The model makes use of NLCD aliases that were previously developed by
Stroud and their consultants.  To incorporate additional data needed for
a simple water quality routine, these aliases need to be mapped back to
their official 2011 NLCD class.  In this process I referenced the
original Scala implementation and the class legend here:

http://www.mrlc.gov/nlcd11_leg.php

It has been noted that we are missing values for 12 & 52, which could be
problematic when we process national NLCD raster data.