Change input format.
The modification types "bmp" and "reclassification" have been unified
into a single type: "change".

For a modification, the word "change" is given as a key with an
associated value which is a string of the form "soil:landuse:bmp". Any
missing items (for example, soil and landuse when a BMP is being
represented) can simply be left blank (for example, "::green_roof").

This is in service of WikiWatershed/model-my-watershed#585