Make sure runoff is non-zero
Since the runoff for BMPs is calculated by subtracting the sum of a
fixed infiltration value read from a table and the evapotranspiration
from the precipitation, it is possible to get a negative runoff.  This
patch addresses that making the runoff the maximum of zero and the
difference just mentioned.

Connects WikiWatershed/model-my-watershed#895