D-Type Infiltration Values for BMPs
The infiltration amounts for Porous Paving, Rain Gardens, and
Infiltration Trenches on soil D have been set to 1/3 of those on soil C.

Because the Rain Garden is actually a mixture of 80% high-intensity
residential and 20% ideal rain garden, it is not generally the case that
the ratios of the infiltration values are 1:3 on D- versus C-type soil.
The exact ratio may also fail to occur when the infiltration value has
been reduced (as when the precipitation is low).

Connects #34