Updates to TR55 & Pitts as recommended by Stroud
After evaluating the model execution in the MMW application, Stroud
provided adjustments to the model and it's inputs.  Tests were updated
accordingly and some removed.  The removed tests were testing that the
threshold of values was relatively inline with some previous expected
output of the model, but due to these changes, that threshold proved to
be larger than reasonable to assert in the tests.