Modified test modules for new curve numbers
In this branch, I updated the curve numbers for all urban land types.
They are now a blend between "pasture in medium condition" and "paved
parking" instead of the default "pasture in good condition" used by the
standard TR55 documentation.  The new curve numbers are calculated using
20% impervious (paved parking) for developed_open, 38% for
developed_low, 65% for developed_med, and 85% for developed_high.  These
percentages are based on those from the standard land uses for WinSLAMM
and the small storm hydrology model.  Because the testing module was
looking for the original curve numbers, I had to modify it to pass.