Rewrote Pitt Model
Rewrote the way the Small Storm Hydrology (Pitt) model was implemented.
Now the model uses data from the Standard Land Uses (SLUs) documented on
the WinSLAMM web page.  The new implementation will calculate runoff up
to 5" of rain.  The runoff is now also dependent on soil type.  The
final runoff for a cell will always be the higher of TR-55 or SSH.  In
all cases, TR-55 will give a higher result before the 5" max rainfall
allowed by the SSH model, allowing for a smooth transition between the
two models.

Note that for this to work I added numpy to the requirements for the
repo!