Renamed infiltration trench to basin
Changed the naming of the BMP in the code to more closely match the UI
and what the BMP is more commonly called.  There is a lot of
inconsistency in the literature wrt what is called a trench or basin and
what constitutes vegetation, but more often trench is used for something
long and skinny and filled with gravel and basin is used for something
more square-ish and fully open.  We are implementing the latter, so I
re-named it basin.